# api/openapi.yaml
openapi: 3.0.3
info:
  title: Investudy API
  version: 0.1.0
  description: |
    REST API веб-ОПиУ. Единственный источник правды для контракта: Go-сервер
    (oapi-codegen) и TS-типы фронта (openapi-typescript) генерируются отсюда.

    Суммы — десятичные строки ("-1234567.89"): NUMERIC(17,2) упирается в
    границу точности Number в JavaScript.

    Аутентификация — cookie `session` (HttpOnly, SameSite=Lax), same-origin
    через nginx. Решения — docs/plans/pnl-architecture.md.
security:
  - cookieAuth: []
paths:
  /api/auth/login:
    post:
      operationId: login
      summary: Вход по логину и паролю
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: |
            Вход выполнен, cookie сессии установлена. Тело — то же, что отдаёт
            `GET /api/me`: после входа фронту не нужен второй запрос.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: Слишком много неудачных попыток
          headers:
            Retry-After:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
...
  responses:
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Нет сессии или она погашена
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        problems:
          description: Нарушения проверки, по одному на пункт.
          type: array
          items:
            type: string
...
    Role:
      type: string
      enum: [division_head, coo, cco, founder, cfo]
...
    Ref:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
    LoginRequest:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
        password:
          type: string
    Me:
      type: object
      required: [id, role, is_admin, has_report]
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/Role'
        division:
          allOf:
            - $ref: '#/components/schemas/Ref'
          nullable: true
        is_admin:
          type: boolean
        has_report:
          description: |
            Отчёт доступен. false — роль не назначена или неизвестна (например,
            viewer из старой таблицы бота): last_line тогда ничего не значит.
          type: boolean
        last_line:
          description: Код последней видимой строки ОПиУ, включительно; null — все.
          type: string
          nullable: true
...